MAR-3160. Refactored store modules. Added store helper. Fix start animation of separator. minor fixes
