+++ client/components/core/Footer/FooterNavbarColumn.vue
@@ -1,7 +1,7 @@
 <template>
   <div
     class="footer-section"
-    :class="`footer-section-${name}`"
+    :class="`footer-section--${name}`"
   >
     <div class="footer-section__content">
       <div class="footer-section__menus">
@@ -46,6 +46,3 @@ export default {
 
 }
 </script>
-
-<style lang="scss" scoped>
-</style>
